lastgenre: Move fetching to client module
and tunelog helper to a utils module since we will need it everywhere.
